misc/HtmlTestSuite/results/test-0429.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<link href="dgrogan@chromium.org" rel="author" title="David Grogan"/>
<link href="https://github.com/w3c/csswg-drafts/issues/3336" rel="help"/>
<meta content="Should cell's percent width be ignored when its table is nested in a shrink to fit block?" name="assert"/>
</head><body><p>Edge 44.17763 and Chrome 70 make this 300px wide. FF makes it 150px wide.</p>
<div data-expected-width="300" id="stf" style="position:absolute; background:blue;">
    <table cellpadding="0" cellspacing="0">
        <tbody><tr><td style="width:50%;">
            <div style="width:150px; height:150px;"/>
        </td>
    </tr></tbody></table>
</div>
<script>
checkLayout('#stf');
</script>
</body></html>
